Made output for table2.
Expanded on the ad hoc crosstabs in the 'Data management' section to produce a CSV with event counts by exposure for both O/O and obese outcomes.

Also reorganized the directory structure to add output/ and to move fig/ therein. Hand-modified table 1 and outputted CSV of table 2 (plus a nicer hand-modified version) added to output/.