Recalculated table 2 stats for obesity outcome.
Realized while updating the obesity analysis script (and incorporating it into the repository) that the follow-up cap will impact each outcome separately. As such, the statistics for table 2 need to be calculated separately. Reproduced various output after making this correction.

This commit represents all of the status of work prior to the Tuesday, May 3 meeting.