Make tables of crude results for breed analyses.
Make tables of crude results for the effect of neutering on O/O for a
breed-specific analysis involving Golden Retrievers (with one table
for all Large breed dogs, one table for Golden Retrievers, and one
table for Large breed dogs excluding Golden Retrievers). This was
inspired by discussion at the 20220607 meeting. See email to Valerie
and Joyce 20220607-1325 for further discussion.